Validate HTML in example tests
