--- run_tests.sh
@@ -1,74 +1,0 @@
-echo TestHelloWorldApp
-python -m unittest test.test_examples.test_examples_app.TestHelloWorldApp
-
-echo TestTemplateApp
-python -m unittest test.test_examples.test_examples_app.TestTemplateApp
-
-echo TestAppendAndRemoveWidgetsApp
-python -m unittest test.test_examples.test_examples_app.TestAppendAndRemoveWidgetsApp
-
-echo TestCloseableApp
-python -m unittest test.test_examples.test_examples_app.TestCloseableApp
-
-echo TestGaugeApp
-python -m unittest test.test_examples.test_examples_app.TestGaugeApp
-
-echo TestGridLayoutApp
-python -m unittest test.test_examples.test_examples_app.TestGridLayoutApp
-
-echo TestLayoutApp
-python -m unittest test.test_examples.test_examples_app.TestLayoutApp
-
-echo TestMatplotlibApp
-python -m unittest test.test_examples.test_examples_app.TestMatplotlibApp
-
-# echo TestMinefieldApp
-# python -m unittest test.test_examples.test_examples_app.TestMinefieldApp
-
-echo TestNotificationApp
-python -m unittest test.test_examples.test_examples_app.TestNotificationApp
-
-echo TestOncloseWindowApp
-python -m unittest test.test_examples.test_examples_app.TestOncloseWindowApp
-
-echo TestPageInternalsApp
-python -m unittest test.test_examples.test_examples_app.TestPageInternalsApp
-
-echo TestPilApp
-python -m unittest test.test_examples.test_examples_app.TestPilApp
-
-echo TestResizablePanes
-python -m unittest test.test_examples.test_examples_app.TestResizablePanes
-
-echo TestResourcesApp
-python -m unittest test.test_examples.test_examples_app.TestResourcesApp
-
-echo TestRootWidgetChangeApp
-python -m unittest test.test_examples.test_examples_app.TestRootWidgetChangeApp
-
-# echo TestSessionApp
-# python -m unittest test.test_examples.test_examples_app.TestSessionApp
-
-echo TestStandaloneApp
-python -m unittest test.test_examples.test_examples_app.TestStandaloneApp
-
-# echo TestSvgplotApp
-# python -m unittest test.test_examples.test_examples_app.TestSvgplotApp
-
-echo TestTabboxApp
-python -m unittest test.test_examples.test_examples_app.TestTabboxApp
-
-echo TestTableWidgetApp
-python -m unittest test.test_examples.test_examples_app.TestTableWidgetApp
-
-echo TestTemplateAdvancedApp
-python -m unittest test.test_examples.test_examples_app.TestTemplateAdvancedApp
-
-# echo TestThreadedApp
-# python -m unittest test.test_examples.test_examples_app.TestThreadedApp
-
-echo TestWebAPIApp
-python -m unittest test.test_examples.test_examples_app.TestWebAPIApp
-
-# echo TestWidgetOverviewApp
-# python -m unittest test.test_examples.test_examples_app.TestWidgetOverviewApp
