+++ test/test_widget.py
@@ -3,24 +3,9 @@
 import unittest
 import remi.gui as gui
 try:
-    from html.parser import HTMLParser
-except ImportError:
-    from HTMLParser import HTMLParser
-
-class SimpleParser(HTMLParser):
-    def __init__(self):
-        HTMLParser.__init__(self)
-        self.elements = []
-
-    def handle_starttag(self, tag, attrs):
-        self.elements.append((tag, dict(attrs)))
-
-def assertValidHTML(text):
-    h = SimpleParser()
-    h.feed(text) 
-    # throws expections if invalid.
-    return True
-
+    from html_validator import SimpleParser, assertValidHTML
+except ValueError:
+    from .html_validator import SimpleParser, assertValidHTML
 
 # tests start here
 class TestTag(unittest.TestCase):
